requirements.txt
pandas
streamlit
faiss-cpu
sentence_transformers